refactor: refactor `DateTimeComponent`
